Add files via upload
The computer will pick a random number between 0 and 50, inclusive.  The user will guess a number, and the computer will tell you if it is too high, too low, or correct.  If it is correct, it will display how many guesses it took you, and ask if you want to play again.