Add files via upload
Overview:  Create a Tic-Tac-Toe application that uses some of the simple components of a GUI.

Interface:  The playing field should consist of nine buttons laid out in a 3x3 grid. You should also have a label indicating your name and the number of wins you have, as well as a label indicating the name and number of wins the second player has.   Two text fields should allow the names of the players to change. Include a “Change name” button, which takes whatever is entered from the text fields and alters the names of the players. If no text is entered, leave the name the way it is. (20 points)

Implementation:  When you press a button in the grid, it should change the label to be an “X” (and disable the button).  The second player should then place an “O” somewhere in the grid (and disable the button).  If anyone wins (or there is a tie), it should update the appropriate score and clear the button labels.