Add files via upload
Overview:  Create a simulation of “Conway’s Life”.  This takes place on a square grid of at least 25 by 25.  All adjacent cells (including diagonal) are considered neighbors.

 

Rules:  

If alive

0 or 1 neighbors - dies of loneliness

2 or 3 neighbors - survives to next round

4 or more neighbors - dies of overcrowding

If dead

3 neighbors - becomes populated

 

Every cell checks its neighbors and calculates whether it will be alive in the next generation.  All cells are then changed, and the calculation occurs again.

 

Interface:  The program should have a JPanel that is the drawing surface. You should be able to click on an individual cell and change it from dead to alive and vice versa. There should be three buttons in addition to the panel. The Step button should show the next generation and stop. (The general algorithm is listed in the overview.) Start should continually step through the generations and display them.  Stop should halt a running simulation.  

 

Implementation:  You should use a two dimensional array to keep track of the current state  - with boundary conditions (the edges of the grid).