Add files via upload
You are currently in the 1-20 lab at Sunset.  All the cool people are hanging out here.

 

There are exits:

WEST

 

There are no items in the room.

 

Then, if you type in WEST:

 

You are in one hall.  You see a computer lab to the east.  Wait; are they playing with robots in there?

 

There are exits:

NORTH SOUTH EAST

 

There is a MOP here.

 

There must be at least 15 rooms, each of which tells you the exits. 

 

There must be at least 5 items. The rooms should tell you what items are in the room. 

 

Items must be able to be picked up and dropped.  The rooms should keep track of items (the items should disappear from the room when picked up or add them when dropped), and the player should have an inventory command. 

 

The game must have a winning condition, in addition to the quit command.