Create CPAP_demo_cscript
In order to create a comprehensive model that is applicable to the scope of this research, an Arduino Thermostat Code was written using the programming language C. Testing for the (mysensor) message was done through an OpenHab software to automate the temperature within a room/household. The C script thoroughly examines the following factors for climate: AC/cooling temperature, ideal temperature (default is the average temperature in Angola)  48℉, and, since the DHT22 sensor is connected to the 4 Channel 24V Relay Module, humidification is balanced simultaneously. Though the C script cannot be tested accurately, a manual is presented below for the different settings/interfaces within the CAD 3D model.