IAMPAIRFILTER
NOTE: CODE should be ran via. "Jupiter Notebook." To simulate an effective, realistic patient-to-filter interface, the following CPAP machine was abided and tested thoroughly: Philips Respironics general CPAP machine was tested on. The following research questions were prompted: a.) Percentage of dust (big chunks of debris (Scale: 6 grams of beans and 10 grams of beans) that had stayed stuck inside of a tube compared to all different kinds of the air filter. b.) The absorbance of different substances such as thicker liquids (water vs. coffee) to simulate snot or just excessive moisture. (This was formulated via the thickness of snot (molecular formula: C102H151O39N31)  vs. the absorbancy of water H₂O) c.) Micro/nanomaterials that are almost invisible and their ability to get caught on an air filter.